start_sample bug in new segments
When check_for_new_segment() is called, this changes fixes a bug where the start_sample variable of the internal state of the channel isn't reset.  This cases the index and RED block start_sample values to keep increasing, when they should be reset to zero for each new segment.  Append_mef_channel_data() was working properly and is not affected by this bug fix.