Bug fix to update_mefd_file()
Instead of using MEF_FULL_FILE_NAME_BYTES as the buffer size to write to the optional .mefd file, just use 1024 instead.  This is how it was before the defintion of MEF_FULL_FILE_NAME_BYTES was changed to be OS dependent.  The code now guarantees each entry in the file will be 1024 bytes, as intended.  The MEF_FULL_FILE_NAME_BYTES was always intended to be an internal constant, meaning it isn't physically manifested in a MEF file.  So a new .h constant should be introduced - this will be done in a future update.